requirements.txt
anyio==3.5.0; python_full_version >= "3.6.2" and python_version >= "3.6"
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0")
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "3.11"
compose==1.2.8
google-api-core==2.7.2; python_version >= "3.6" and python_version < "3.11"
google-auth==2.6.6; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
google-cloud-bigquery-storage==2.13.1; python_version >= "3.6" and python_version < "3.11"
google-cloud-bigquery==3.0.1; python_version >= "3.6" and python_version < "3.11"
google-cloud-core==2.3.0; python_version >= "3.6" and python_version < "3.11"
google-cloud-tasks==2.8.1; python_version >= "3.6"
google-crc32c==1.3.0; python_version >= "3.6" and python_version < "3.11"
google-resumable-media==2.3.2; python_version >= "3.6" and python_version < "3.11"
googleapis-common-protos==1.56.0; python_version >= "3.6" and python_version < "3.11"
grpc-google-iam-v1==0.12.4; python_version >= "3.6"
grpcio-status==1.45.0; python_version >= "3.6" and python_version < "3.11"
grpcio==1.45.0; python_version >= "3.6" and python_version < "3.11"
h11==0.12.0; python_version >= "3.6"
httpcore==0.14.7; python_version >= "3.6"
httpx==0.22.0; python_version >= "3.6"
idna==3.3; python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.2"
numpy==1.22.3; python_version >= "3.8" and python_version < "3.11"
packaging==21.3; python_version >= "3.6" and python_version < "3.11"
proto-plus==1.20.3; python_version >= "3.6" and python_version < "3.11"
protobuf==3.20.1; python_version >= "3.7" and python_version < "3.11"
pyarrow==7.0.0; python_version >= "3.7" and python_version < "3.11"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0") or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0")
pyparsing==3.0.8; python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.8"
python-dateutil==2.8.2; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.3.0"
requests==2.27.1; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
rfc3986==1.5.0; python_version >= "3.6"
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0")
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
sniffio==1.2.0; python_full_version >= "3.6.2" and python_version >= "3.6"
urllib3==1.26.9; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_full_version >= "3.6.0" and python_version < "3.11" and python_version >= "3.6"
